Change how results are compared to fastfilters v1
It is practically impossible to obtain results that are precisely equal
to the old implementation without a loss in performance and loss in the
implementation flexibility. Therefore some amount of numerical
differences between implementations is allowed.

Previously, we tried to compare maximum absolute differences between
filter outputs, using the full uint8 dynamic range (mapped to float32).
The reasoning was to be as close to the real-world usage as possible.
This resulted in significant differences between implementations.

However, the old implementation used much laxer tests in order to verify
its outputs w.r.t. to the Vigra library. First, it used values in the
[0; 1) range, where precision is much higher. Second, it used mean
absolute difference (*not* max) for 3D multi-channel filters, where
numerical differences are especially pronounced.

Now, we are using the same strategy for the compatibility testing. The
tests are a bit different: we test for more values of sigma (scale), but
skip testing the structure tensor with the different derivative scale
(because it is almost always derived from the main outer scale).

Finally, the script `resultcompare.py` has been updated to match the
test code.